implement combine function for creating global models and link parameters
- add new tests set
